# Specalize NewTuple to support 0 to 6 direct arguments.