--- a/compiler/util.py
+++ b/compiler/util.py
@@ -26,6 +26,12 @@ _SIMPLE_CHARS = set(string.digits + string.letters + string.punctuation + " ")
 _ESCAPES = {'\t': r'\t', '\r': r'\r', '\n': r'\n', '"': r'\"', '\\': r'\\'}
 
 
+# This is the max length of a direct allocation tuple supported by the runtime.
+# This should match the number of specializations found in
+# runtime/tuple_direct.go.
+MAX_DIRECT_TUPLE = 6
+
+
 class ParseError(Exception):
 
   def __init__(self, node, msg):