--- a/runtime/float.go
+++ b/runtime/float.go
@@ -84,7 +84,7 @@ func floatGetNewArgs(f *Frame, args Args, _ KWArgs) (*Object, *BaseException) {
 	if raised := checkMethodArgs(f, "__getnewargs__", args, FloatType); raised != nil {
 		return nil, raised
 	}
-	return NewTuple(args[0]).ToObject(), nil
+	return NewTuple1(args[0]).ToObject(), nil
 }
 
 func floatGT(f *Frame, v, w *Object) (*Object, *BaseException) {