--- a/runtime/int.go
+++ b/runtime/int.go
@@ -102,7 +102,7 @@ func intGetNewArgs(f *Frame, args Args, _ KWArgs) (*Object, *BaseException) {
 	if raised := checkMethodArgs(f, "__getnewargs__", args, IntType); raised != nil {
 		return nil, raised
 	}
-	return NewTuple(args[0]).ToObject(), nil
+	return NewTuple1(args[0]).ToObject(), nil
 }
 
 func intGT(f *Frame, v, w *Object) (*Object, *BaseException) {