--- a/runtime/long.go
+++ b/runtime/long.go
@@ -124,7 +124,7 @@ func longGetNewArgs(f *Frame, args Args, _ KWArgs) (*Object, *BaseException) {
 	if raised := checkMethodArgs(f, "__getnewargs__", args, LongType); raised != nil {
 		return nil, raised
 	}
-	return NewTuple(args[0]).ToObject(), nil
+	return NewTuple1(args[0]).ToObject(), nil
 }
 
 func longGT(x, y *big.Int) bool {