This eliminates an extra allocation for the backing array by storing it in the memory directly following the tuple itself.

In the micro benchmarks, the big impact is on the dict iter items implementation where 2-tuples are generated for every entry in the dict. The 6 element test seemed to have high variation when run, but the rest paints a reasonable picture.

```
name                       old time/op    new time/op    delta
DictIterItems/0-elements      518ns ± 0%     517ns ± 1%     ~     (p=0.540 n=4+5)
DictIterItems/1-elements      629ns ± 4%     577ns ± 0%   -8.24%  (p=0.008 n=5+5)
DictIterItems/2-elements      687ns ± 0%     646ns ± 0%   -5.97%  (p=0.016 n=4+5)
DictIterItems/3-elements      807ns ± 1%     720ns ± 1%  -10.83%  (p=0.008 n=5+5)
DictIterItems/4-elements      902ns ± 4%     779ns ± 0%  -13.61%  (p=0.008 n=5+5)
DictIterItems/5-elements     1.02µs ± 1%    0.87µs ± 2%  -15.04%  (p=0.008 n=5+5)
DictIterItems/6-elements     1.29µs ± 1%    1.21µs ±14%     ~     (p=0.151 n=5+5)
DictIterItems/7-elements     1.39µs ± 1%    1.20µs ± 2%  -13.55%  (p=0.008 n=5+5)
DictIterItems/8-elements     1.45µs ± 2%    1.28µs ± 2%  -11.94%  (p=0.008 n=5+5)

name                       old alloc/op   new alloc/op   delta
DictIterItems/0-elements       320B ± 0%      320B ± 0%     ~     (all equal)
DictIterItems/1-elements       384B ± 0%      384B ± 0%     ~     (all equal)
DictIterItems/2-elements       448B ± 0%      448B ± 0%     ~     (all equal)
DictIterItems/3-elements       512B ± 0%      512B ± 0%     ~     (all equal)
DictIterItems/4-elements       576B ± 0%      576B ± 0%     ~     (all equal)
DictIterItems/5-elements       640B ± 0%      640B ± 0%     ~     (all equal)
DictIterItems/6-elements       704B ± 0%      704B ± 0%     ~     (all equal)
DictIterItems/7-elements       768B ± 0%      768B ± 0%     ~     (all equal)
DictIterItems/8-elements       832B ± 0%      832B ± 0%     ~     (all equal)

name                       old allocs/op  new allocs/op  delta
DictIterItems/0-elements       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
DictIterItems/1-elements       8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.008 n=5+5)
DictIterItems/2-elements       10.0 ± 0%       8.0 ± 0%  -20.00%  (p=0.008 n=5+5)
DictIterItems/3-elements       12.0 ± 0%       9.0 ± 0%  -25.00%  (p=0.008 n=5+5)
DictIterItems/4-elements       14.0 ± 0%      10.0 ± 0%  -28.57%  (p=0.008 n=5+5)
DictIterItems/5-elements       16.0 ± 0%      11.0 ± 0%  -31.25%  (p=0.008 n=5+5)
DictIterItems/6-elements       18.0 ± 0%      12.0 ± 0%  -33.33%  (p=0.008 n=5+5)
DictIterItems/7-elements       20.0 ± 0%      13.0 ± 0%  -35.00%  (p=0.008 n=5+5)
DictIterItems/8-elements       22.0 ± 0%      14.0 ± 0%  -36.36%  (p=0.008 n=5+5)
```

The changes to the compiler do have an impact on the output code. In most cases, it actually reduces the number of temporary variables needed at any given time. While I doubt that will have any impact on the actual compiled output (given the coming SSA backend), less allocations should have an impact. I can try grabbing numbers if anyone thinks it would be interesting, but I'm not sure we have any good benchmarks which would be affected by this change.