@nairb774 2017-01-26
I can the methods into tuple.go. I had them there initially, but it was a lot of visual clutter in my opinion.